refactor into plog
